chore(monitors): Add components folder
