+++ static/app/views/monitors/components/onboarding.tsx
